accounts-summariser/pyproject.toml
[project]
name = "accounts-summariser"
version = "0.1.0"
description = "An ADK agent that summarises FTSE company accounts"
authors = [
	{name="Roland Dunn", email="roland.dunn@gmail.com"}
]
license = "Apache License 2.0"
readme = "README.md"
packages = [{ include = "accounts_summariser", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
google-adk = { version = ">=0.5.0", extras = ["eval"] }
google-cloud-aiplatform = {extras = ["adk", "agent-engines"], version = "^1.88.0"}
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
